+++ src/trefoil/catalog.cr
@@ -6,5 +6,5 @@ require "./thread"
 
 module Trefoil
   alias Catalog = Array(Thread)
-  alias CatalogResponse = Array(NamedTuple(page: Int8, threads: Array(PostInfo)))
+  alias CatalogResponse = Array(NamedTuple(page: Int32, threads: Array(PostInfo)))
 end
